MBS-11502 extract jsonrpcclient to package
