+++ subprojects/test-runner/report-viewer/src/main/kotlin/com/avito/jsonrpc/JsonRpcClient.kt
@@ -1,7 +1,6 @@
-package com.avito.reportviewer.internal
+package com.avito.jsonrpc
 
 import com.avito.http.internal.RequestMetadata
-import com.avito.reportviewer.internal.model.RfcRpcRequest
 import com.github.salomonbrys.kotson.fromJson
 import com.google.gson.Gson
 import okhttp3.MediaType.Companion.toMediaType
