+++ subprojects/test-runner/report-viewer/src/main/kotlin/com/avito/jsonrpc/RpcResult.kt
@@ -1,4 +1,4 @@
-package com.avito.reportviewer.internal.model
+package com.avito.jsonrpc
 
 /**
  * jsonrpc and id params omitted
