+++ subprojects/test-runner/report-viewer/src/main/kotlin/com/avito/reportviewer/ReportsApi.kt
@@ -27,6 +27,9 @@ public interface ReportsApi : ReportsAddApi, ReportsFetchApi {
      * @param reportId идентификатор запуска в отчетах (run_id)
      * @param analyzerKey имя анализатора
      * @param preparedData произвольные данные анализатора
+     *
+     * todo remove JsonElement from public api
+     * todo rewrite Russian
      */
     public fun pushPreparedData(reportId: String, analyzerKey: String, preparedData: JsonElement): Result<Unit>
 }
