+++ subprojects/test-runner/report-viewer/src/main/kotlin/com/avito/reportviewer/ReportsApiFactory.kt
@@ -2,30 +2,30 @@ package com.avito.reportviewer
 
 import com.avito.http.HttpClientProvider
 import com.avito.http.RetryInterceptor
+import com.avito.jsonrpc.JsonRpcClient
 import com.avito.logger.LoggerFactory
 import com.avito.report.serialize.createReportGson
-import com.avito.reportviewer.internal.JsonRpcClient
 import com.avito.reportviewer.internal.ReportsApiImpl
 import java.util.concurrent.TimeUnit
 
 public object ReportsApiFactory {
 
-    private const val TIMEOUT_SEC = 10L
+    private const val DEFAULT_TIMEOUT_SEC = 10L
 
     public fun create(
         host: String,
         httpClientProvider: HttpClientProvider,
         loggerFactory: LoggerFactory,
-        retryRequests: Boolean = true
+        retryRequests: Boolean = true,
+        readWriteTimeoutSec: Long = DEFAULT_TIMEOUT_SEC
     ): ReportsApi {
         return ReportsApiImpl(
             loggerFactory = loggerFactory,
             client = JsonRpcClient(
                 host = host,
                 httpClient = httpClientProvider.provide()
-                    .connectTimeout(TIMEOUT_SEC, TimeUnit.SECONDS)
-                    .writeTimeout(TIMEOUT_SEC, TimeUnit.SECONDS)
-                    .readTimeout(TIMEOUT_SEC, TimeUnit.SECONDS)
+                    .writeTimeout(readWriteTimeoutSec, TimeUnit.SECONDS)
+                    .readTimeout(readWriteTimeoutSec, TimeUnit.SECONDS)
                     .apply {
                         if (retryRequests) {
                             addInterceptor(
