+++ subprojects/test-runner/report-viewer/src/main/kotlin/com/avito/reportviewer/internal/model/CthulhuTestCase.kt
@@ -4,4 +4,4 @@ package com.avito.reportviewer.internal.model
  * Ctulhu is an Avito's Test Model System
  * prepared_data field. After analysis matches with TMS if has a test_case_id
  */
-internal data class CtulhuTestCase(val description: String?)
+internal data class CthulhuTestCase(val description: String?)
