+++ subprojects/test-runner/report-viewer/src/main/kotlin/com/avito/reportviewer/internal/model/ListResult.kt
@@ -32,5 +32,5 @@ internal data class ListResult(
         get() = testName.substringAfter(testNameDelimiter)
 }
 
-// gson не переварит private поле
+// open as internal for gson
 internal const val testNameDelimiter = "::"
