+++ subprojects/test-runner/report-viewer/src/main/kotlin/com/avito/reportviewer/internal/model/PreparedData.kt
@@ -9,7 +9,7 @@ internal data class PreparedData(
     @SerializedName("tc_build") val tcBuild: String?,
     @SerializedName("skip_reason") val skipReason: String?,
     @SerializedName("external_id") val externalId: String?,
-    @SerializedName("testcase") val ctulhuTestCase: CtulhuTestCase?,
+    @SerializedName("testcase") val cthulhuTestCase: CthulhuTestCase?,
     @SerializedName("features") val features: List<String>?,
     @SerializedName("tag_id") val tagId: List<Int>?,
     @SerializedName("feature_id") val featureIds: List<Int>?,
