+++ subprojects/test-runner/report-viewer/src/main/kotlin/com/avito/reportviewer/model/Report.kt
@@ -2,6 +2,8 @@ package com.avito.reportviewer.model
 
 /**
  * Один прогон тестов по координатам
+ *
+ * todo internal
  */
 public data class Report(
     val id: String,
