+++ subprojects/test-runner/report-viewer/src/main/kotlin/com/avito/reportviewer/model/ReportCoordinates.kt
@@ -4,6 +4,8 @@ import java.io.Serializable
 
 /**
  * Один уникальный запуск тест сьюта
+ *
+ * todo internal
  */
 public data class ReportCoordinates(
     val planSlug: String,
