+++ subprojects/test-runner/report-viewer/src/main/kotlin/com/avito/reportviewer/model/RunId.kt
@@ -7,6 +7,8 @@ package com.avito.reportviewer.model
  *          should be the same if source not changed, so commit hash should be used instead of buildId
  *          local builds could use current time, to avoid git interaction
  * @param buildTypeId is used to distinguish runs with the same API on `develop` branch
+ *
+ * todo internal
  */
 public data class RunId(
     private val prefix: String? = null,
