--- subprojects/test-runner/report-viewer/src/main/kotlin/com/avito/reportviewer/model/TestStaticDataExt.kt
@@ -1,20 +1,0 @@
-package com.avito.reportviewer.model
-
-import com.avito.report.model.TestStaticDataPackage
-import com.avito.test.model.DeviceName
-
-public fun TestStaticDataPackage.Companion.fromSimpleRunTest(simpleRunTest: SimpleRunTest): TestStaticDataPackage =
-    TestStaticDataPackage(
-        name = simpleRunTest.name,
-        device = DeviceName(simpleRunTest.deviceName),
-        description = simpleRunTest.description,
-        testCaseId = simpleRunTest.testCaseId,
-        dataSetNumber = simpleRunTest.dataSetNumber,
-        externalId = simpleRunTest.externalId,
-        featureIds = simpleRunTest.featureIds,
-        tagIds = simpleRunTest.tagIds,
-        priority = simpleRunTest.priority,
-        behavior = simpleRunTest.behavior,
-        kind = simpleRunTest.kind,
-        flakiness = simpleRunTest.flakiness
-    )
