+++ subprojects/test-runner/report-viewer/src/main/resources/META-INF/proguard/report-viewer.pro
@@ -1,5 +1,2 @@
--keep class com.avito.android.test.** { *; }
--keep interface com.avito.android.test.** { *; }
-
--keep class com.avito.report.model.** { *; }
--keep interface com.avito.report.model.** { *; }
+-keep class com.avito.reportviewer.model.** { *; }
+-keep interface com.avito.reportviewer.model.** { *; }
